Finfin (#205)
* otc fix (#195)

* otc fix

* otc fix

* Finfin (#196)

* otc fix

* otc fix

* otc fix

* Finfin (#198)

* otc fix

* otc fix

* otc fix

* otc fix

* seeder fix

* seeder fix

* seeder fix

* seeder fix

* Future contract fix (#200)

* otc fix

* otc fix

* otc fix

* otc fix

* seeder fix

* seeder fix

* seeder fix

* seeder fix

* futures fix

* Seeder fix (#201)

* otc fix

* otc fix

* otc fix

* otc fix

* seeder fix

* seeder fix

* seeder fix

* seeder fix

* futures fix

* futures fix

* Seeder fix (#202)

* otc fix

* otc fix

* otc fix

* otc fix

* seeder fix

* seeder fix

* seeder fix

* seeder fix

* futures fix

* futures fix

* seeders

---------

+++ berza-servis/src/main/java/rs/edu/raf/order/model/StranaPonudaDTO.java
@@ -15,5 +15,5 @@ public class StranaPonudaDTO {
 
     private Integer quantity;
 
-    private BigDecimal amountOffered;
+    private Double amountOffered;
 }
